move utils to a utils directory
