+++ src/project/baseComponent.tsx
@@ -10,7 +10,7 @@
  */
 
 import "./index.scss"
-import type {CognitoUser} from "../cognito-utils"
+import type {CognitoUser} from "../utils/cognito"
 import Prism from "prismjs"
 import React from "react"
 import cache from "../cache"
