+++ src/projects/index.tsx
@@ -12,13 +12,13 @@
 import "./index.scss"
 import {Link, useParams} from "react-router-dom"
 import {Project, isProject} from "../project/baseComponent"
-import type {CognitoUser} from "../cognito-utils"
+import type {CognitoUser} from "../utils/cognito"
 import DefaultPhoto from "../images/default.svg"
 import {Img} from "../elements"
 import React from "react"
 import {Spinner} from "../bootstrap"
 import UserContext from "../userContext"
-import {arrayToChunks} from "../utils"
+import {arrayToChunks} from "../utils/misc"
 import cache from "../cache"
 import handleError from "../errorHandler"
 import notify from "../notify"
